+++ Change-log.md
@@ -5,6 +5,7 @@
 * Send ping only if WebSocket is connected.
 * Concept of splitLargeFrames flag.
   Working towards [ng2-stompjs#120](https://github.com/stomp-js/ng2-stompjs/issues/120)
+* Concept of forceBinaryWSFrames.
 
 # 5.1.0 (2018-12-16)